requirements.txt
asgiref==3.3.1
case-converter==1.0.2
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
djangorestframework=3.12.2
idna==2.10
lxml==4.6.1
mysqlclient==2.0.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-json-logger==2.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toolz==0.11.1
trading-calendars==2.0.0
urllib3==1.26.2
